requirements.txt
asgiref==3.4.1
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
DateTime==4.3
django==3.2
django-debug-toolbar==3.2.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
idna==2.10
mkl-fft==1.3.0
mkl-random==1.2.1
mkl-service==2.3.0
numpy==1.20.2
olefile==0.46
Pillow==8.3.1
psycopg2==2.8.4
pycparser==2.20
pycryptodome==3.8.2
pyOpenSSL==19.1.0
PySocks==1.7.1
pytz==2021.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.25.9
zope.interface==4.7.1